Login com senha criptografada concluido, proximo passo: criar cadastro cliente e sessao
