+++ mocks/credentials.dir
@@ -1,0 +1,1 @@
+'joao', (0, 98)
