+++ models/Resultado.py
@@ -1,0 +1,5 @@
+class resultado:
+            
+            def __init__(self,resultado,mensagem):
+                self.resultado = resultado
+                self.mensagem = mensagem
